+++ dart_sample/lib/sample_routes.dart
@@ -28,6 +28,7 @@ class SampleRoutes {
   static const String linearProgressIndicator = '/linear-progress-indicator';
   static const String circularProgressIndicator =
       '/circular-progress-indicator';
+  static const String slider = '/slider';
   static const String card = '/card';
   static const String listTile = '/list-tile';
   static const String floatingActionButton = '/floating-action-button';