+++ src/Plumix.Material/ThemeData.cs
@@ -158,6 +158,7 @@ public sealed record ThemeData
     private CheckboxThemeData? _checkboxTheme;
     private SwitchThemeData? _switchTheme;
     private RadioThemeData? _radioTheme;
+    private SliderThemeData? _sliderTheme;
 
     public ThemeData(
         TargetPlatform? platform = null,
@@ -207,7 +208,8 @@ public sealed record ThemeData
         ProgressIndicatorThemeData? progressIndicatorTheme = null,
         CheckboxThemeData? checkboxTheme = null,
         SwitchThemeData? switchTheme = null,
-        RadioThemeData? radioTheme = null)
+        RadioThemeData? radioTheme = null,
+        SliderThemeData? sliderTheme = null)
     {
         Platform = platform ?? ResolveDefaultPlatform();
         Brightness = brightness ?? Brightness.Light;
@@ -257,6 +259,7 @@ public sealed record ThemeData
         _checkboxTheme = checkboxTheme;
         _switchTheme = switchTheme;
         _radioTheme = radioTheme;
+        _sliderTheme = sliderTheme;
     }
 
     public TargetPlatform Platform { get; init; }
@@ -417,6 +420,12 @@ public sealed record ThemeData
     {
         get => _radioTheme ?? new RadioThemeData();
         init => _radioTheme = value;
+    }
+
+    public SliderThemeData SliderTheme
+    {
+        get => _sliderTheme ?? new SliderThemeData();
+        init => _sliderTheme = value;
     }
 
     public static ThemeData Light { get; } = new();